renamed SharedComponents Products to FeaturedProducts
