+++ src/pages/Home/Home.js
@@ -1,6 +1,6 @@
 import React from "react"
 import WelcomeSection from "./WelcomeSection"
-import Products from "../../SharedComponents/Products"
+import FeaturedProducts from "../../SharedComponents/FeaturedProducts"
 import PageWrapper from "../../SharedComponents/PageWrapper"
 
 import About from "./About"
@@ -12,7 +12,7 @@ import RecipeSection from './RecipeSection'
 const Home = () =>
     <PageWrapper>
       <WelcomeSection />
-      <Products title={"Featured Products"} hasTopBottomBorders={true}/>
+      <FeaturedProducts title={"Featured Products"} hasTopBottomBorders={true}/>
       <About />
       <PhotoSection />
       <Process />
