+++ src/pages/Recipe/Recipe.js
@@ -5,7 +5,7 @@ import styled from "styled-components";
 
 import { device } from "../../utils/devices";
 
-import Products from '../../SharedComponents/Products';
+import FeaturedProducts from '../../SharedComponents/FeaturedProducts';
 import PageWrapper from '../../SharedComponents/PageWrapper';
 import StyledH1 from '../../SharedComponents/StyledH1';
 
@@ -88,7 +88,7 @@ const Recipe = ({
   return (
     <PageWrapper>
       {createRecipe()}
-      <Products title={"Explore the Shop"} />
+      <FeaturedProducts title={"Explore the Shop"} />
     </PageWrapper>
   );
 };
